requirement.txt
APScheduler==3.6.3
asgiref==3.2.7
Django==3.0.5
paho-mqtt==1.5.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
uWSGI==2.0.18
